started cleaning/refactoring lucys code
