+++ app/src/main/java/com/mobileproto/hireddit/hireddit/reddit/CommentCallback.java
@@ -1,4 +1,4 @@
-package com.mobileproto.hireddit.hireddit;
+package com.mobileproto.hireddit.hireddit.reddit;
 
 import java.util.ArrayList;
 
@@ -6,5 +6,5 @@ import java.util.ArrayList;
  * Created by lwilcox on 11/5/2015.
  */
 public interface CommentCallback {
-    public void callback(ArrayList<String> commentList);
+    void callback(ArrayList<String> commentList);
 }
