+++ app/src/main/java/com/mobileproto/hireddit/hireddit/visuals/MainFragment.java
@@ -7,7 +7,6 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
-import android.widget.TextView;
 
 import com.mobileproto.hireddit.hireddit.R;
